requirements.txt
Flask~=3.1.2
python-dotenv~=1.1.1
requests~=2.32.5
numpy~=2.3.3